start & setup menu almost working
